Updated np.float32 to float

+++ pose_pipeline/wrappers/deep_sort_yolov4/deep_sort/detection.py
@@ -27,10 +27,10 @@ class Detection(object):
     """
 
     def __init__(self, tlwh, confidence, cls, feature):
-        self.tlwh = np.asarray(tlwh, dtype=np.float)
+        self.tlwh = np.asarray(tlwh, dtype=float)
         self.confidence = float(confidence)
         self.cls = cls
-        self.feature = np.asarray(feature, dtype=np.float32)
+        self.feature = np.asarray(feature, dtype=float)
 
     def to_tlbr(self):
         """Convert bounding box to format `(min x, min y, max x, max y)`, i.e.,
